+++ app/domain/template_generation/__init__.py
@@ -1,0 +1,6 @@
+from .errors import TemplateError, TemplateGenerationError
+
+__all__ = [
+    "TemplateError",
+    "TemplateGenerationError",
+]