+++ app/web/__init__.py
@@ -6,6 +6,7 @@ from .shift_reports.routes import shift_report_details_ns, shift_report_ns
 from .project_materials.routes import project_material_ns
 from .materials.routes import material_ns
 from .project_works.routes import project_work_ns
+from .template_generation.routes import template_ns
 from .work_categories.routes import work_category_ns
 from .work_material_relations.routes import work_material_relation_ns
 from .work_prices.routes import work_price_ns
@@ -26,3 +27,4 @@ def register_namespaces(api: Api):
     api.add_namespace(shift_report_material_ns)
     api.add_namespace(work_price_ns)
     api.add_namespace(work_ns)
+    api.add_namespace(template_ns)